Move "sl_quick_actions" entry command to a top level file
Our reloader does not replace commands which are just imported (for
example in `_init.py`) to the top level.  So we stay true and move the
command to an extra file at the top level.
