+++ _init.py
@@ -1,2 +1,1 @@
 from .lint.generic_text_command import sl_generic_text_cmd  # noqa: F401
-from .lint.quick_fix import sl_quick_actions  # noqa: F401
