+++ lint/quick_fix.py
@@ -4,12 +4,11 @@ from itertools import chain
 import re
 
 import sublime
-import sublime_plugin
 
 from . import persist
-from . import util
 from .generic_text_command import replace_view_content, text_command
 flatten = chain.from_iterable
+
 
 MYPY = False
 if MYPY:
@@ -41,82 +40,6 @@ class QuickAction:
     def description(self):
         # type: () -> str
         return "   ".join(filter(None, (self.subject, self.detail)))
-
-
-class sl_quick_actions(sublime_plugin.TextCommand):
-    def is_enabled(self, quiet=False):
-        # type: (bool) -> bool
-        if quiet:
-            return len(self.view.sel()) == 1
-        else:
-            return True
-
-    def run(self, edit, quiet=False):
-        # type: (sublime.Edit, bool) -> None
-        view = self.view
-        window = view.window()
-        assert window
-
-        if not quiet:
-            if len(self.view.sel()) != 1:
-                window.status_message("Not implemented for multiple selections")
-                return
-
-        sel = view.sel()[0]
-        filename = util.get_filename(view)
-        if sel.empty():
-            char_selection = sublime.Region(sel.a, sel.a + 1)
-            errors = get_errors_where(
-                filename,
-                lambda region: region.intersects(char_selection)
-            )
-            if not errors:
-                sel = view.full_line(sel.a)
-                errors = get_errors_where(
-                    filename,
-                    lambda region: region.intersects(sel)
-                )
-        else:
-            errors = get_errors_where(
-                filename,
-                lambda region: region.intersects(sel)
-            )
-
-        def on_done(idx):
-            # type: (int) -> None
-            if idx < 0:
-                return
-
-            action = actions[idx]
-            apply_fix(action.fn, view)
-
-        actions = sorted(
-            list(actions_for_errors(errors, view)),
-            key=lambda action: (-len(action.solves), action.description)
-        )
-        if not actions:
-            if quiet:
-                window.show_quick_panel(
-                    ["No actions available."],
-                    lambda x: None
-                )
-            else:
-                window.status_message("No actions available")
-        elif len(actions) == 1:
-            on_done(0)
-        else:
-            window.show_quick_panel(
-                [action.description for action in actions],
-                on_done
-            )
-
-
-def get_errors_where(filename, fn):
-    # type: (str, Callable[[sublime.Region], bool]) -> List[LintError]
-    return [
-        error for error in persist.file_errors[filename]
-        if fn(error['region'])
-    ]
 
 
 def actions_for_errors(errors, view=None):
